requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
shapely>=2.0.0
scipy>=1.11.0
pandas>=2.0.0
pyproj>=3.5.0
tqdm>=4.65.0
pyyaml>=6.0
# Oriented R-CNN (optional – only needed for train_oriented_rcnn.py)
# mmengine>=0.10.0
# mmcv>=2.1.0
# mmdet>=3.2.0
# mmrotate>=1.0.0rc1
matplotlib>=3.7.0
scikit-learn>=1.3.0
pytest>=7.4.0
